Make minichemiters have the event too
